% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_perseo.R
\name{.extract_contrasts}
\alias{.extract_contrasts}
\title{Extract contrast tibble from a PERSEO result object}
\usage{
.extract_contrasts(x)
}
\description{
Extract contrast tibble from a PERSEO result object
}
\keyword{internal}
