% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_perseo.R
\name{.extract_counts}
\alias{.extract_counts}
\title{Extract counts matrix from a PERSEO result object (for MA plot)}
\usage{
.extract_counts(x)
}
\description{
Extract counts matrix from a PERSEO result object (for MA plot)
}
\keyword{internal}
