% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_perseo.R
\docType{data}
\name{.perseo_sig_colours}
\alias{.perseo_sig_colours}
\title{PERSEO colour palette for significance categories}
\format{
An object of class \code{character} of length 4.
}
\usage{
.perseo_sig_colours
}
\description{
PERSEO colour palette for significance categories
}
\keyword{internal}
