% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_transformations.R
\name{family_groups}
\alias{family_groups}
\title{Family groups by theoretical support}
\usage{
family_groups()
}
\value{
List with character vectors of families by support: count, unit, positive, real.
}
\description{
Family groups by theoretical support
}
