% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlss_fitting.R
\name{gof_score}
\alias{gof_score}
\title{Filliben correlation of the normalized quantile residuals (higher = better fit).}
\usage{
gof_score(fit)
}
\description{
Filliben correlation of the normalized quantile residuals (higher = better fit).
}
\keyword{internal}
