% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_perseo.R
\name{print.perseo_results}
\alias{print.perseo_results}
\title{Print method for PERSEO results}
\usage{
\method{print}{perseo_results}(x, ...)
}
\arguments{
\item{x}{A `perseo_results` object returned by `run_perseo()`.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for PERSEO results
}
