% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlss_fitting.R
\name{select_by_ic_gof}
\alias{select_by_ic_gof}
\title{Index of the minimum-IC family, breaking near-ties (IC within `delta`) by best GoF.}
\usage{
select_by_ic_gof(ic_values, gof_values, delta = 2)
}
\description{
Index of the minimum-IC family, breaking near-ties (IC within `delta`) by best GoF.
}
\keyword{internal}
