% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_transformations.R
\name{infer_support}
\alias{infer_support}
\title{Infer empirical support of a response vector}
\usage{
infer_support(y)
}
\arguments{
\item{y}{Numeric vector.}
}
\value{
One of "count", "unit", "positive", "real", or "none".
}
\description{
Infer empirical support of a response vector
}
